requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
psycopg2
httpx
python-dotenv